+++ cap/config/deploy.rb
@@ -1,8 +1,8 @@
 # config valid for current version and patch releases of Capistrano
 lock "~> 3.11.0"
 
-set :application, "<application name>"
-set :repo_url, "<repo>"
+set :application, APP_CONFIG["PROJECT_NAME"]
+set :repo_url, APP_CONFIG["GIT_REPO"]
 
 # set to deploy only what is in the src directory
 set :repo_tree, 'src/'
@@ -14,8 +14,6 @@ set :keep_releases, 5
 set :magento_deploy_chmod_d, 711
 set :magento_deploy_chmod_f, 644
 
-# add wp folder to linked dirs
-append :linked_dirs, 'wp'
 
-set :magento_auth_public_key, '<you pulic_key here>'
-set :magento_auth_private_key, '<your private_key here'
+set :magento_auth_public_key, APP_CONFIG["MAGENTO_AUTH_PUBLIC_KEY"]
+set :magento_auth_private_key, APP_CONFIG["MAGENTO_AUTH_PRIVATE_KEY"]