+++ cap/config/deploy/staging-sample.rb
@@ -1,3 +1,0 @@
-server '<server>', user: '<user>', roles: %w{app db web}
-set :deploy_to, '<path>'
-set :branch, proc { `git rev-parse --abbrev-ref <branch>`.chomp }