+++ docker/tasks/build.sh
@@ -1,4 +1,3 @@
 #!/usr/bin/env bash
-cd ../
-source ../../.env
+PROJECT_NAME=$1
 docker-compose -p $PROJECT_NAME build