+++ docker/tasks/db-import.sh
@@ -1,11 +1,11 @@
 #!/usr/bin/env bash
-cd ../
-source ../../.env
+PROJECT_NAME=$1
+FILE=$2
 
-if [ -z "$1" ]
+if [ -z "$2" ]
   then
     echo "enter sql file name"
 else
-    zcat mysqldump/$@ | docker-compose -p $PROJECT_NAME exec -T -u root db mysql -umagento -pmagento magento
+    zcat mysqldump/$FILE | docker-compose -p $PROJECT_NAME exec -T -u root db mysql -umagento -pmagento magento
 	docker-compose -p $PROJECT_NAME exec -T -u root db mysql -umagento -pmagento magento --execute="update core_config_data set value='http://localhost/' where path='web/unsecure/base_url';"
 fi