+++ docker/tasks/db-shell.sh
@@ -1,5 +1,3 @@
 #!/usr/bin/env bash
-cd ../
-source ../../.env
-
+PROJECT_NAME=$1
 docker-compose -p $PROJECT_NAME exec --user root db bash