+++ docker/tasks/kill.sh
@@ -1,6 +1,4 @@
 #!/usr/bin/env bash
-cd ../
-source ../../.env
-
+PROJECT_NAME=$1
 docker-compose -p $PROJECT_NAME stop
 docker-compose -p $PROJECT_NAME down -v