+++ tasks/magento.sh
@@ -1,6 +1,5 @@
 #!/usr/bin/env bash
-source ../../.env
-cd ../docker
-
-docker-compose -p $PROJECT_NAME exec --user www-data magento /var/www/html/bin/magento $@
+PROJECT_NAME=$1
+TASK=$2
+docker-compose -p $PROJECT_NAME exec --user www-data magento /var/www/html/bin/magento $TASK
 exit 0